src/requirements_dev.txt
bandit==1.6.2
django-debug-toolbar==2.1
django-silk==3.0.4
pytest==5.3.1
pytest-django==3.7.0
pytest-flake8==1.0.4
pytest-forked==1.1.3
pytest-cov==2.8.1
pytest-xdist==1.30.0
requests==2.22.0
responses==0.10.8
safety==1.8.5
# TODO: add monkeytype and pyannotate
